Removes the combinat dependency by enumerating partitions ourselves.
The combinat package was used to enumerate partitions, in order to
generate all orbits of a set. However, this package doesn't build
with recent GHC (and I had troubles with building this package
in the past as well), so I wanted to get rid of it. It was only
used for a single feature.

As a bonus, the new enumeration is slightly more efficient (only
a constant factor). But this is hardly noticeable, because most
of the work happens elsewhere.